feat(permission): 添加权限实体类
- 定义权限实体类继承基础实体类
- 配置数据库表名为 sys_permission
- 添加菜单权限相关字段包括父级ID、菜单标题、路径、组件等
- 集成 Excel 导入导出功能注解
- 添加 Swagger API 文档注解支持
- 实现菜单权限管理模块基本信息定义

+++ backend/src/main/java/com/lawoffice/system/entity/Permission.java
@@ -58,7 +58,7 @@ public class Permission extends BaseEntity {
 
     @ExcelProperty("菜单排序")
     @Schema(description = "菜单排序")
-    private Double sortNo;
+    private Integer sortNo;
 
     @ExcelIgnore
     @Schema(description = "聚合子路由: 1是0否")
